Generalize functions to perform multiple range requests at once.
This involves modifying downloadIndexRange so that it can perform multiple
range requests at once for efficiency. Right now, we achieve this via
parallelization, but it can easily be modified to use a single multi-part
request if the server supports it.

We also have to rename a lot of the functions as we're no longer dealing
with a single gene or set in many of them.