Cleaned up names in docs so that 'indices' refer only to integers.
No more talk about "Gesel indices", which are now referred to as "Gesel
database files". We also refer to the string gene identifiers as "gene names"
to distinguish them better from the gene indices.

We convert the output of fetchAllGenes into a data frame to make it easier
to describe what the gene indices are. This requires a little fix in the line
decompression code to avoid removal of a trailing empty line.