Centralized cache handling so that it can be easily flushed.
This is primarily intended for testing but can also be useful to trigger
a re-download of all data assets when the remote files change. This
effectively replaces the use.preloaded=FALSE option.

We also added more caching in the *Some* functions so we can reduce the
number of requests by leveraging data from previous calls.