Pass cache and fetch= functions via a configuration object.
This was motivated by the realization that users could change the fetch
arguments to point to different databases, but gesel functions would
store everything in the same cache. Now, the cache is passed as an
explicit argument that corresponds to a specific set of fetch functions.

This change also makes it easier to call gesel functions with different
fetch functions, because we can just pass a single configuration object
rather than having to remember to pass all of the fetch= arguments.